As stated in the previous commit, I worked on a different copy of the project. That means non of my fixes were committed.
This fixes it. Here's the text of the commit:

- Changed the way of storing NPCs from a Collection Object to a SQL DB as this
is more convenient for.. pretty much anything
- Added SQL table for NPCs, Comments and Header Files (Spells are coming too)
- Implemented Header and Comment Add, Delete, Edit and View options in the
menubar (NPCs are upcoming next, of course)

NOTES:
- I did not change the way of NPCs being saved at this point, that will be done
soon enough. Therefore the Collection code is still in the file. This will be
removed after I fixed the SQL for NPCs.
- SQL Queries do not check for errors properly. I totally forgot to do so. I
will implement that later, but it's not high on my to-do list.